SYNC

+++ src/movement.py
@@ -14,8 +14,8 @@ def linear(angle=0, speed=0):
     def move(obj, pos=(0, 0), time=0):
         time = float(time) / 1000
         distance = speed * time
-        dx = int(round(distance * angle_cos))
-        dy = int(round(distance * angle_sin))
+        dx = round(distance * angle_cos)
+        dy = round(distance * angle_sin)
         x, y = pos
         return x + dx, y - dy
